Separate message to new component
