+++ WebClient/src/store/Chat/selectors.ts
@@ -2,8 +2,14 @@ import { ChatTypeEnum } from '../../api';
 import { AppState } from '../store';
 import { AnnouncementMessage, DefaultMessage, MessageInStore, MessageType } from './store';
 
+export interface DefaultMessageWithUserName extends DefaultMessage {
+    userName?: string;
+}
+
+export type MessageInStoreWithUserName = (DefaultMessageWithUserName | AnnouncementMessage);
+
 // FIXME: add re-select, because this selector causes constant chat rerenders
-export const chatMessages = (chatType: ChatTypeEnum) => (state: AppState) => {
+export const chatMessages = (chatType: ChatTypeEnum) => (state: AppState): MessageInStoreWithUserName[] | undefined => {
     if (!state.currentUser.user?.roomId)
         return undefined;
